revert (rest of) r365382 by adrianS. if it still crashes on 64-bit big endian, i want a proper bug report and not some crap-tastic commit.
svn path=/trunk/KDE/kdebase/workspace/; revision=730298

+++ kdm/backend/dm.h
@@ -200,7 +200,7 @@ typedef struct CfgDep {
 
 typedef struct CfgArr {
 	void *data;     /* config value array; allocated */
-	long *idx;      /* config index array; alias */
+	int *idx;      /* config index array; alias */
 	CfgDep dep;     /* filestamp */
 	int numCfgEnt;  /* number of config entries */
 } CfgArr;
